+++ src/App/Vivianne.Wpf/Controls/DoubleValue.cs
@@ -1,0 +1,14 @@
+﻿using TheXDS.MCART.Types.Base;
+
+namespace TheXDS.Vivianne.Controls;
+
+public class DoubleValue : NotifyPropertyChanged
+{
+    private double _Value;
+
+    public double Value
+    {
+        get => _Value;
+        set => Change(ref _Value, value);
+    }
+}
